refactor out utils
